fixes #15040 - remove user fragment cache initializer
The initializer added in c3e0fed caused the foreman initializer to fail
as loaded the User model, which it assumes meant the table was present.
This isn't the case when db:migrate runs for the first time.

The new initializer has been removed so as to only keep the existing
cache expiry in the foreman initializer, which now checks for presence
of the table instead. The existing initializer is also compatible with
memcache, preventing an exception starting Rails.