Fixes #24712 - Fix performance regression on API
This commit reverts the code part of 01809199 which has been found to
cause significant performance regressions. The tests are left in place.
It seems that the tests pass even without the change done in the code,
perhaps some other change between the time the original issue was
reported and the fix was introduced led to the fix being redundant.